requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.2
cffi==1.16.0
cryptography==41.0.7
greenlet==3.0.1
importlib-metadata==7.0.1
pycparser==2.21
pygal==3.0.4
PyMySQL==1.0.2
python-dotenv==1.0.0
SQLAlchemy==1.4.46
typing_extensions==4.8.0
zipp==3.17.0
